add working inverse kinematics
