+++ elir_inverse_kinematics/setup.py
@@ -1,0 +1,9 @@
+from distutils.core import setup
+from catkin_pkg.python_setup import generate_distutils_setup
+
+d = generate_distutils_setup(
+    packages=['kinematics_solver'],
+    package_dir={'': 'src'}
+)
+
+setup(**d)
