+++ elir_move_group/src/move_group_python.py
@@ -1,0 +1,158 @@
+#!/usr/bin/env python
+
+# Software License Agreement (BSD License)
+#
+# Copyright (c) 2013, SRI International
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions
+# are met:
+#
+#  * Redistributions of source code must retain the above copyright
+#    notice, this list of conditions and the following disclaimer.
+#  * Redistributions in binary form must reproduce the above
+#    copyright notice, this list of conditions and the following
+#    disclaimer in the documentation and/or other materials provided
+#    with the distribution.
+#  * Neither the name of SRI International nor the names of its
+#    contributors may be used to endorse or promote products derived
+#    from this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+# FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+# COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+# INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+# BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+# LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+# ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+# POSSIBILITY OF SUCH DAMAGE.
+#
+# Author: Acorn Pooley, Mike Lautman
+
+import sys
+import copy
+import rospy
+import moveit_commander
+import moveit_msgs.msg
+import geometry_msgs.msg
+from math import pi
+from std_msgs.msg import String
+from moveit_commander.conversions import pose_to_list
+from elir_move_group.srv import MoveRequest
+from kinematics_solver.solver import elir_inverse_kinematics
+
+## END_SUB_TUTORIAL
+
+def all_close(goal, actual, tolerance):
+  """
+  Convenience method for testing if a list of values are within a tolerance of their counterparts in another list
+  @param: goal       A list of floats, a Pose or a PoseStamped
+  @param: actual     A list of floats, a Pose or a PoseStamped
+  @param: tolerance  A float
+  @returns: bool
+  """
+  all_equal = True
+  if type(goal) is list:
+    for index in range(len(goal)):
+      if abs(actual[index] - goal[index]) > tolerance:
+        return False
+
+  elif type(goal) is geometry_msgs.msg.PoseStamped:
+    return all_close(goal.pose, actual.pose, tolerance)
+
+  elif type(goal) is geometry_msgs.msg.Pose:
+    return all_close(pose_to_list(goal), pose_to_list(actual), tolerance)
+
+  return True
+
+class MoveGroupElir(object):
+  """MoveGroupPythonIntefaceTutorial"""
+  def __init__(self):
+    super(MoveGroupElir, self).__init__()
+
+    self.move_request_service = rospy.Service('motion_planning/MoveRequest',MoveRequest, self.go_to_pose_goal)
+    ## Instantiate a `RobotCommander`_ object. This object is the outer-level interface to
+    ## the robot:
+    robot = moveit_commander.RobotCommander()
+
+    scene = moveit_commander.PlanningSceneInterface()
+
+    self.group_name = "f_arm"
+    self.group = moveit_commander.MoveGroupCommander(self.group_name)
+
+    display_trajectory_publisher = rospy.Publisher('/move_group/display_planned_path',
+                                                   moveit_msgs.msg.DisplayTrajectory,
+                                                   queue_size=20)
+    self.custom_planner = elir_inverse_kinematics()
+
+    # Misc variables
+    self.robot = robot
+    self.scene = scene
+    self.group = group
+    self.planning_frame = planning_frame
+    self.eef_link = eef_link
+    self.group_names = group_names
+
+  def go_to_pose_goal(self,req):
+    
+    x_goal = req.x
+    z_goal = req.z
+    elbow_up = req.elbow_up
+    self.group_name = req.group
+
+    self.group = moveit_commander.MoveGroupCommander(self.group_name)
+
+    self.custom_planner.inverse_kinematics(x_goal,z_goal,elbow_up)
+    group = self.group
+
+    pose_goal = geometry_msgs.msg.Pose()
+    pose_goal.position.x =  0.611913837579
+    pose_goal.position.y = 0.0000000000000
+    pose_goal.position.z = 0.0859528009998
+    pose_goal.orientation.y =  0.525553383506
+    pose_goal.orientation.w = 0.850760625021
+
+    group.set_pose_target(pose_goal)
+
+    ## Now, we call the planner to compute the plan and execute it.
+    plan = group.go(wait=True)
+    # Calling `stop()` ensures that there is no residual movement
+    group.stop()
+    # It is always good to clear your targets after planning with poses.
+    # Note: there is no equivalent function for clear_joint_value_targets()
+    group.clear_pose_targets()
+
+    ## END_SUB_TUTORIAL
+
+    # For testing:
+    # Note that since this section of code will not be included in the tutorials
+    # we use the class variable rather than the copied state variable
+    current_pose = self.group.get_current_pose().pose
+    return all_close(pose_goal, current_pose, 0.01)
+
+
+def main():
+  moveit_commander.roscpp_initialize(sys.argv)
+  rospy.init_node('elir_move_group', anonymous=True)
+  try:
+    print "============ Press `Enter` to begin the tutorial by setting up the moveit_commander (press ctrl-d to exit) ..."
+    raw_input()
+    tutorial = MoveGroupElir()
+
+    print "============ Press `Enter` to execute a movement using a pose goal ..."
+    raw_input()
+    tutorial.go_to_pose_goal()
+
+    print "============ Python tutorial demo complete!"
+  except rospy.ROSInterruptException:
+    return
+  except KeyboardInterrupt:
+    return
+
+if __name__ == '__main__':
+  main()
